+++ apps/ui-service/src/app/(platform)/products/debttrack/dashboard/page.tsx
@@ -6,12 +6,17 @@ import { Debt } from "@/shared/constants/types"
 import { AddEntityCard, EntityCard } from "@/shared/components/entity-card"
 import { EntityType } from "@/shared/components/entity-card/data"
 import StatCardStack from "@/shared/components/stat-card/stat-card-stack"
+import { useUserContext } from "@/context/user.provider"
 
 export default function Page() {
+  const [{ searchKeyword }] = useUserContext()
   const debts = useQuery<Debt[]>({
-    queryKey: ["get-debts"],
-    queryUrl: endPoints.debt,
+    queryKey: ["get-debts", searchKeyword],
+    queryUrl: `${endPoints.debt}?searchKeyword=${encodeURIComponent(
+      searchKeyword
+    )}`,
     method: HTTPMethods.GET,
+    suspense: !!searchKeyword.length ? false : true,
   })
 
   const renderDebts = debts?.data?.map((debt) => (