+++ apps/ui-service/src/app/(platform)/products/wealthanalyzer/dashboard/page.tsx
@@ -6,12 +6,17 @@ import { Institution } from "@/shared/constants/types"
 import { AddEntityCard, EntityCard } from "@/shared/components/entity-card"
 import { EntityType } from "@/shared/components/entity-card/data"
 import StatCardStack from "@/shared/components/stat-card/stat-card-stack"
+import { useUserContext } from "@/context/user.provider"
 
 export default function Page() {
+  const [{ searchKeyword }] = useUserContext()
   const institutions = useQuery<Institution[]>({
-    queryKey: ["get-institutions"],
-    queryUrl: endPoints.institution,
+    queryKey: ["get-institutions", searchKeyword],
+    queryUrl: `${endPoints.institution}?searchKeyword=${encodeURIComponent(
+      searchKeyword
+    )}`,
     method: HTTPMethods.GET,
+    suspense: !!searchKeyword.length ? false : true,
   })
 
   const renderInstitutions = institutions?.data?.map((institution) => (